Fix cart display issues
